Moved asset related functions to a new file
